backend/requirements.txt
# --- Web framework & server ---
fastapi
uvicorn[standard]

# --- Database & migrations ---
sqlalchemy
psycopg2-binary          # PostgreSQL driver (safe for dev)
alembic                  # optional but recommended for migrations

# --- Data validation & environment ---
pydantic
python-dotenv

# --- Auth & security ---
python-jose[cryptography]
passlib[bcrypt]

# --- Google Gemini (LLM) ---
google-generativeai
httpx                     # required for async transport

# --- LangGraph workflow orchestration ---
langgraph
langchain-core

# --- Document generation ---
python-docx
python-pptx

# --- Utilities ---
typing-extensions
